Finished initial implementation of updated CPlayerHealth.cs script.
Health is now based on states, and includes a downed state for the player, in addition to both alive and dead.

Refactored various bits of code to work with the new health script. Minor alterations.

Added a reference check to prevent a null reference exception in CThirdPersonAnimController.cs.

Removed an unused debug.log method from CAudioSystem.cs.